refactor(routes): decompose ui.py and api.py into domain sub-packages
Split the two monolithic route files (ui.py: 1726 lines, api.py: 1112 lines)
into focused domain modules under agentception/routes/ui/ and
agentception/routes/api/ respectively.

ui/ modules: overview, agents, telemetry, dag, config, ab_testing,
  brain_dump, roles_ui, github_ui, transcripts, worktrees, docs,
  api_reference, templates_ui
api/ modules: pipeline, control, config, intelligence, telemetry,
  worktrees, issues

Shared helpers (_TEMPLATES singleton, _find_agent, formatters) are
isolated in ui/_shared.py and re-exported from ui/__init__.py so
existing imports from agentception.routes.ui continue to resolve.
api/_shared.py holds _SENTINEL, _build_agent_task, and related helpers.

app.py import paths are unchanged. Verified with mypy (0 errors) and
curl smoke tests against the running service.
